backend/requirements.txt
  Flask
  rembg
  Pillow
  numpy
  